improvement: generate the form_post interstitial on upgrade for OAuth apps
Extracts the interstitial generation into `AshAuthentication.Phoenix.Igniter`
(shared by the installer and upgrader), detecting OAuth/OIDC strategies across
all resources. The upgrader generates it for existing apps that have an
OAuth/OIDC strategy; idempotent and a no-op otherwise.
